Refactor fetching
